Conflicts:
	src/main/java/pl/psnc/dl/wf4ever/model/Builder.java
	src/test/java/pl/psnc/dl/wf4ever/model/BaseTest.java

+++ src/main/java/pl/psnc/dl/wf4ever/oauth/UserResource.java
@@ -1,6 +1,7 @@
 package pl.psnc.dl.wf4ever.oauth;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.sql.SQLException;
@@ -25,17 +26,15 @@ import org.apache.log4j.Logger;
 import org.openrdf.rio.RDFFormat;
 
 import pl.psnc.dl.wf4ever.auth.RequestAttribute;
-import pl.psnc.dl.wf4ever.auth.UserCredentials;
-import pl.psnc.dl.wf4ever.common.db.UserProfile;
 import pl.psnc.dl.wf4ever.connection.DigitalLibraryFactory;
 import pl.psnc.dl.wf4ever.connection.SemanticMetadataServiceFactory;
+import pl.psnc.dl.wf4ever.db.UserProfile;
+import pl.psnc.dl.wf4ever.db.dao.UserProfileDAO;
 import pl.psnc.dl.wf4ever.dl.ConflictException;
 import pl.psnc.dl.wf4ever.dl.DigitalLibraryException;
 import pl.psnc.dl.wf4ever.dl.UserMetadata;
 import pl.psnc.dl.wf4ever.model.Builder;
 import pl.psnc.dl.wf4ever.model.RO.ResearchObject;
-import pl.psnc.dl.wf4ever.rosrs.ROSRService;
-import pl.psnc.dl.wf4ever.sms.QueryResult;
 
 import com.sun.jersey.api.NotFoundException;
 
@@ -124,11 +123,15 @@ public class UserResource {
     private Response getUser(@PathParam("U_ID") String urlSafeUserId, RDFFormat rdfFormat)
             throws DigitalLibraryException {
         String userId = new String(Base64.decodeBase64(urlSafeUserId));
-
-        QueryResult qs = ROSRService.SMS.get().getUser(UserProfile.generateAbsoluteURI(null, userId), rdfFormat);
+        UserProfileDAO userProfileDAO = new UserProfileDAO();
+        UserProfile user = userProfileDAO.findByLogin(userId);
+        if (user == null) {
+            throw new NotFoundException("User not found");
+        }
+        InputStream userDesc = user.getAsInputStream(rdfFormat);
 
         if (DigitalLibraryFactory.getDigitalLibrary().userExists(userId)) {
-            return Response.ok(qs.getInputStream()).type(qs.getFormat().getDefaultMIMEType()).build();
+            return Response.ok(userDesc).type(rdfFormat.getDefaultMIMEType()).build();
         } else {
             return Response.status(Status.NOT_FOUND).type("text/plain").entity("User " + userId + " does not exist")
                     .build();
@@ -176,8 +179,6 @@ public class UserResource {
         String password = UUID.randomUUID().toString().replaceAll("-", "").substring(0, 20);
         boolean created = DigitalLibraryFactory.getDigitalLibrary().createUser(userId, password,
             username != null && !username.isEmpty() ? username : userId);
-        UserCredentials creds = new UserCredentials(userId, password);
-        creds.save();
         UserMetadata user = DigitalLibraryFactory.getDigitalLibrary().getUserProfile(userId);
         SemanticMetadataServiceFactory.getService(user).close();
 
@@ -214,19 +215,16 @@ public class UserResource {
             throws DigitalLibraryException, NotFoundException, ClassNotFoundException, IOException, NamingException,
             SQLException, pl.psnc.dl.wf4ever.dl.NotFoundException {
         String userId = new String(Base64.decodeBase64(urlSafeUserId));
-        Set<URI> list = ROSRService.SMS.get().findResearchObjectsByCreator(
-            UserProfile.generateAbsoluteURI(null, userId));
-        for (URI uri : list) {
-            ResearchObject ro = ResearchObject.get(builder, uri);
+        UserProfileDAO userProfileDAO = new UserProfileDAO();
+        UserProfile user = userProfileDAO.findByLogin(userId);
+        if (user == null) {
+            throw new NotFoundException("User not found");
+        }
+        Set<ResearchObject> ros = ResearchObject.getAll(builder, user);
+        for (ResearchObject ro : ros) {
             ro.delete();
         }
-
-        ROSRService.SMS.get().removeUser(URI.create(userId));
         DigitalLibraryFactory.getDigitalLibrary().deleteUser(userId);
-        UserCredentials creds = UserCredentials.findByUserId(userId);
-        if (creds == null) {
-            throw new NotFoundException();
-        }
-        creds.delete();
+        userProfileDAO.delete(user);
     }
 }
